Framingham score calculation migrated
Framingham CHD: Instead of looking for Age in the output, I used the rule fired("calculate age") precondition for some of the rules.
Remarks: This implementation might need some simplification in the future, now there are 50+ rules. And in addition, it does not handle mmol/l values.
The CHD assessment did not need to much modification.