Glasgow Imrie Evaluation migrated
Guidelines migrated. Only small changes were needed (output -> itput, event time)
Note 1:
The guideline does what it is in description, but the shown probabilities are strange (in fact they are the  same in the referenced article, but for example for score 7, there is only 1 case). I would suggest to remove probability information, because it might be missleading, having a 54-years old patient with all the lab values wrong, and saying 0% chance of pancreatis. Even if there is a note that score >3 indicates high risk.
Note 2:
Guideline accepts only specific units for eaach value (although in the architypes there might be more). This is not specified in the description.
Note 3:
Decimals don't work in the execution tab (but does in the test tab), but I think this is a known issue.