HEART score
Guideline migrated. Changes: output <-input, event time.
Tests work as it is described in the guideline.
Note: I was not happy that a patient with a myocardits and arrythmias can be classified as patient that should be discharged while this can be a life-threatening condition otherwise difficult to discover, as troponin is a very good and specific indicator of heart muscle damage, I think it would be better to change slightly the recommendations and description that this guideline rules out MI, stroke and other MACE, not all life-threatening conditions, and scoring low does not automatically mean discharging from the ED.