ABG migrated
Changes:
(1)Event time,
(2)in delta ratio assessment rule, unit comparison was removed (as it was not set before, therefore making no sense, and it hindered the rule to run; in addition, the correct unit would not be "1");
(3)in set cAG rule a new precondition: "anion gap" exists is added; the guideline runs without it, but can very easily to run to problem, if one of the ions are not set.

Further comments:
(1) in "Set AG if metabolic acidosis: No AG" rule, I see no reason, why AG is limited to 11-14 interval, instead of just keeping the upper bound (14), and allowing everything below. I didn't find anything in the reference literature either.
(2) Misspelling: "compensatory respiratory acidosis"

All rules are tested.