Asthma predictive score migrated
Changes:
(1)Small changes: event time, outp-> inp
(2) in Set minor criteria count rule I added a precondition requiring the existence of eosinophilia, otherwise the whole guideline won't run in case of no eosinophilia.
Note:
I don't see the reason, why eosinophils are given in the same unit as WBC-s, as it is rare that a patient's eos count can reach 1*10^9, it is not very easy to test the lack of eosinophilia with realistic values. For me, even 0.1 was not executed even in the test tab, 0.5 however worked (of course, in the execution tab only 1 works, so many test features has this value, with corresponding unrealistic WBC count)