MEWS migrated
Changes:
(1) Event time
(2) In MEWS.v2 temperature had to be referenced by its magnitude in "Set temperature points to 2" rule, otherwise, it was impossible to even read the file. It took some time though to find the source of error. In fact, the "°" sign is was messing up with the editor.