THRIVE migrated
THRIVE migrated. Although I think in the current form THRIVE score assessment makes no sense, as it fires only if the score is either 0 or 9.
In addition, it contradicts the literature, where it says: "The THRIVE score strongly predicts outcomes among patients undergoing endovascular stroke treatment (Fig 2). Patients with a low THRIVE score of 0–2 had good outcomes in 64.7% of cases, patients with a moderate THRIVE score of 3–5 had good outcomes in 43.5% of cases, and patients with a high THRIVE score of 6–9 had good outcomes in 10.6% of cases (Fig 2 A; P < .001, Mantel-Haenszel χ2 test). Patients with a low THRIVE score of 0–2 had a mortality rate of 5.9% by 90 days, patients with a moderate THRIVE score of 3–5 had a mortality rate of 30.1% by 90 days, and patients with a high THRIVE score of 6–9 had a mortality rate of 56.4% by 90 days". If you agree, I would implement rather this.
Current changes:
(1) Event time
(2)  output->input