QRISK2-2015 migrated
QRISK2-2015 migrated, but this is not a complete guideline, so cannot be tested. A basic problem was "-" sign here as well. Several expressions in the rules had to be rewritten, that they work as intended (i.e. a mathematically identical description has to be formulated not using *-x sign. *(-x) expression does not work either. This guideline also shows that the bug with "-" sign is quite critical.
Since the original guideline was unfinished, tests could not have been run.