HAS-BLED
HAS-BLED migrated. Many changes, the original file was impossible to open either.
Changes (I remember):
(1) Event time, output -> input
(2) "'Yes' rule not fired" preconditions in "No" rules, since they were not the opposite. Now a condition is true, if any of the precondition is true. This was the intention of the writer, but this should be discussed with experts in the field, since the guideline doesn't say anything about this.
(3): CR is now compared without unit. Since the editor is unable to parse the "micro" sign.
(4): Age rules modified that they work correctly.