ABIC score for Alcoholic Hepatitis
(1) Note that it was impossible to make 6.71 as a threshold, so 6.7 is now.
(2) In high risk group, the 1 year mortality is smaller than the 90 days mortality, which makes no sense. I was unable to open the original research paper citing this result, but maybe this is due to the small number of patients in the respective group.