Update
In the calcualte NEWS2 score, instead of the temperature the hypercapnic respiratory failure element was used. The hypercapnic respiratory failure element should not be part of the score. Now this is corrected.
Originally, the reason behind the error was, that the rule had to be edited in the plain text (the expression editor otherwise would mess up), and the wrong reference was used.