Adjustments is done based on comments.
Guideline:
Description: The last sentence needs better English improvement.

Purpose: Avoid using definitive language. Instead of saying Diagnoses posttramautic stress say Assists in the diagnosis of....

Archetypes:
Description: Same as the guideline.
Purpose: It's not good to repeat the same purpose between the two archetypes because each one of them fulfills a different role. Instead for the observation I would put something like: "To record information relevant to the DSM-5 assessment". For the evaluation something like: "To record the result of the DSM-5 assessment".