A-a O2 gradient
I deviated from the calculator which is available on MdCalc, as it might not be the most usable. I used the references and the following lecture about A-a gradient: https://www.youtube.com/watch?v=pEsOv_pdOUQ and decided to implement three versions of it:
- a simplified
- a similar to the MdCalc version
- the most general one.
All guidelines include the patient's temperature as input, as vapor pressure depends on it and as nowadays the guideline might be frequently used with pneumonia or ARDS patients where fever is present.